+++ src/WorkspaceProcesses.ts
@@ -24,10 +24,6 @@ export async function UpdateWorkspace(forceCreate: boolean = false) {
 		FileProcesses.GetSubfolders(globals.glistappsPath).map(folder => {
 			if (fs.existsSync(path.join(folder, "CMakeLists.txt"))) {
 				workspaceFolders.push(folder)
-				if (!fs.existsSync(path.join(folder, ".vscode"))) {
-					vscode.window.showInformationMessage("Launch configurations are not found for project named " + path.basename(folder) + ". Creating launch configurations...");
-					fs.cpSync(path.join(extension.extensionPath, 'GlistApp', '.vscode'), path.join(folder, '.vscode'), { recursive: true });
-				}
 			}
 		});
 		workspaceFolders.push(globals.glistEnginePath);
@@ -35,8 +31,10 @@ export async function UpdateWorkspace(forceCreate: boolean = false) {
 			folders: workspaceFolders.map(folder => ({ path: folder })),
 		};
 
+		CheckLaunchConfigurations();
+
 		fs.writeFileSync(globals.workspaceFilePath, JSON.stringify(workspaceContent, null, 2));
-		vscode.window.showInformationMessage('Workspace Updated.');
+		vscode.window.showInformationMessage('Workspace configured.');
 
 		//VS Code will restart if another workspace is active.
 		await vscode.commands.executeCommand('vscode.openFolder', vscode.Uri.file(globals.workspaceFilePath), false);
@@ -47,6 +45,7 @@ export async function UpdateWorkspace(forceCreate: boolean = false) {
 }
 
 export async function AddNewProjectToWorkspace(projectName: string, forceCreate: boolean = false) {
+	CheckLaunchConfigurations();
 	if (!fs.existsSync(globals.workspaceFilePath) || forceCreate) {
 		extension.extensionJsonData.firstRun = false;
 		extension.extensionJsonData.secondRun = true;
@@ -104,6 +103,7 @@ export async function CloseNonExistentFileTabs() {
 	try {
 		const visibleWindows = vscode.window.tabGroups.activeTabGroup.tabs;
 		for (const visibleWindow of visibleWindows) {
+			if(!visibleWindow.input) continue;
 			const filePath = JSON.parse(JSON.stringify(visibleWindow.input, null, 2));
 			if (fs.existsSync(filePath.uri.fsPath)) continue;
 			await vscode.window.tabGroups.close(visibleWindow, false);
@@ -121,4 +121,15 @@ export async function LaunchWorkspace() {
 	else { // If no workspace file found, create one and launch the workspace
 		await UpdateWorkspace(true);
 	}
+}
+
+async function CheckLaunchConfigurations() {
+	FileProcesses.GetSubfolders(globals.glistappsPath).map(folder => {
+		if (fs.existsSync(path.join(folder, "CMakeLists.txt"))) {
+			if (!fs.existsSync(path.join(folder, ".vscode"))) {
+				vscode.window.showInformationMessage("Creating launch configurations for " + path.basename(folder));
+				fs.cpSync(path.join(extension.extensionPath, 'GlistApp', '.vscode'), path.join(folder, '.vscode'), { recursive: true });
+			}
+		}
+	});
 }